chore(html): move types to types module
